@@ -35,17 +35,23 @@ dependencies {
   testRuntimeOnly "org.spekframework.spek2:spek-runner-junit5"
 }
 
+test {
+  useJUnitPlatform {
+    includeEngines "spek2", "junit-vintage", "junit-jupiter"
+  }
+}
+
 compileKotlin {
   kotlinOptions {
-    languageVersion = "1.4"
-    jvmTarget = "11"
+    languageVersion = "1.7"
+    jvmTarget = "17"
   }
 }
 
 compileTestKotlin {
   kotlinOptions {
-    languageVersion = "1.4"
-    jvmTarget = "11"
+    languageVersion = "1.7"
+    jvmTarget = "17"
   }
 }
 
