@@ -38,7 +38,9 @@ public interface CredentialsRepository<T extends Credentials> {
    */
   boolean has(String name);
 
-  /** @return A new set containing all known credentials */
+  /**
+   * @return A new set containing all known credentials
+   */
   Set<T> getAll();
 
   /**
@@ -57,6 +59,8 @@ public interface CredentialsRepository<T extends Credentials> {
    */
   void delete(String name);
 
-  /** @return Type of credentials this repository can store */
+  /**
+   * @return Type of credentials this repository can store
+   */
   String getType();
 }
