@@ -21,7 +21,9 @@ import java.time.Duration;
 import java.util.function.Supplier;
 
 public class RetrySupport {
-  /** @deprecated replaced by {@link #retry(Supplier, int, Duration, boolean)} */
+  /**
+   * @deprecated replaced by {@link #retry(Supplier, int, Duration, boolean)}
+   */
   @Deprecated
   public <T> T retry(Supplier<T> fn, int maxRetries, long retryBackoffMillis, boolean exponential) {
     return retry(fn, maxRetries, Duration.ofMillis(retryBackoffMillis), exponential);
