@@ -15,8 +15,7 @@
  */
 
 package com.spinnaker.netflix.kork.plugins.internal
-class PluginJarTest{}
-/*
+
 import com.netflix.spinnaker.kork.plugins.internal.PluginJar
 import com.spinnaker.netflix.kork.plugins.SomeFeatureExtension
 //import com.spinnaker.netflix.kork.plugins.TestPlugin
@@ -30,8 +29,8 @@ import java.util.jar.JarInputStream
 import strikt.api.expect
 import strikt.assertions.isEqualTo
 
-class PluginJarTest : JUnit5Minutests {
-
+class PluginJarTest {//: JUnit5Minutests {
+/*
   fun tests() = rootContext<Fixture> {
     context("PluginJar") {
       fixture {
@@ -64,6 +63,5 @@ class PluginJarTest : JUnit5Minutests {
       .manifestAttribute("Plugin-Requires", pluginRequires)
       .extension(SomeFeatureExtension::class.java.name)
       .build()
-  }
+  }*/
 }
-*/
