@@ -25,9 +25,9 @@ public class UserSecretReferenceTest {
 
   @Test
   public void encodedURIData() {
-    var ref = UserSecretReference.parse(
-      "secret://engine%20identifier?first=hello%20world&second=%68%65%6c%6c%6f%20%77%6f%72%6c%64"
-    );
+    var ref =
+        UserSecretReference.parse(
+            "secret://engine%20identifier?first=hello%20world&second=%68%65%6c%6c%6f%20%77%6f%72%6c%64");
     assertEquals("engine identifier", ref.getEngineIdentifier());
     var parameters = ref.getParameters();
     assertEquals("hello world", parameters.get("first"));
