@@ -46,6 +46,7 @@ import liquibase.resource.ClassLoaderResourceAccessor;
 import org.jooq.DSLContext;
 import org.jooq.Query;
 import org.jooq.SQLDialect;
+import org.jooq.Table;
 import org.jooq.impl.DataSourceConnectionProvider;
 import org.jooq.impl.DefaultConfiguration;
 import org.jooq.impl.DefaultDSLContext;
@@ -63,13 +64,17 @@ public class SqlTestUtil {
 
   public static String tcPgJdbcUrl = "jdbc:tc:postgres:10.13:///test";
 
-  /** @deprecated Please use the TestContainers db initializers where possible, instead of H2 */
+  /**
+   * @deprecated Please use the TestContainers db initializers where possible, instead of H2
+   */
   @Deprecated
   public static TestDatabase initDatabase() {
     return initDatabase("jdbc:h2:mem:test;MODE=MYSQL");
   }
 
-  /** @deprecated Please use the TestContainers db initializers where possible, instead of H2 */
+  /**
+   * @deprecated Please use the TestContainers db initializers where possible, instead of H2
+   */
   @Deprecated
   public static TestDatabase initPreviousDatabase() {
     return initDatabase("jdbc:h2:mem:test_previous;MODE=MYSQL");
@@ -269,7 +274,7 @@ public class SqlTestUtil {
     context.meta().getTables().stream()
         .filter(
             table ->
-                table.getType().isTable()
+                table.getType().isInstance(Table.class)
                     && table.getSchema().getName().equals(schema)
                     && !table.getName().equals(configuration.getDatabaseChangeLogTableName())
                     && !table.getName().equals(configuration.getDatabaseChangeLogLockTableName()))
