@@ -43,12 +43,13 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.core.env.Environment;
 
 @Configuration
-@EnableConfigurationProperties({StackdriverConfig.SpectatorStackdriverConfigurationProperties.class})
+@EnableConfigurationProperties({
+  StackdriverConfig.SpectatorStackdriverConfigurationProperties.class
+})
 @ConditionalOnProperty("spectator.stackdriver.enabled")
 public class StackdriverConfig {
 
-  @Autowired
-  ServerProperties serverProperties;
+  @Autowired ServerProperties serverProperties;
 
   @ConfigurationProperties("spectator")
   public static class SpectatorStackdriverConfigurationProperties {
@@ -148,37 +149,39 @@ public class StackdriverConfig {
   /**
    * Schedule a thread to flush our registry into stackdriver periodically.
    *
-   * <p>
-   * This configures our StackdriverWriter as well.
+   * <p>This configures our StackdriverWriter as well.
    */
   @Bean
-  public StackdriverWriter defaultStackdriverWriter(Environment environment,
-                                                    Registry registry,
-                                                    SpectatorStackdriverConfigurationProperties spectatorStackdriverConfigurationProperties)
-    throws IOException {
+  public StackdriverWriter defaultStackdriverWriter(
+      Environment environment,
+      Registry registry,
+      SpectatorStackdriverConfigurationProperties spectatorStackdriverConfigurationProperties)
+      throws IOException {
     Logger log = LoggerFactory.getLogger("StackdriverConfig");
     log.info("Creating StackdriverWriter.");
-    Predicate<Measurement> filterNotSpring = new Predicate<Measurement>() {
-      public boolean test(Measurement measurement) {
-        // Dont store measurements that dont have tags.
-        // These are from spring; those of interest were replicated in spectator.
-        if (measurement.id().tags().iterator().hasNext()) {
-          return true;
-        }
+    Predicate<Measurement> filterNotSpring =
+        new Predicate<Measurement>() {
+          public boolean test(Measurement measurement) {
+            // Dont store measurements that dont have tags.
+            // These are from spring; those of interest were replicated in spectator.
+            if (measurement.id().tags().iterator().hasNext()) {
+              return true;
+            }
 
-        return false;
-      }
-    };
+            return false;
+          }
+        };
     Predicate<Measurement> measurementFilter;
 
-    final String prototypeFilterPath = spectatorStackdriverConfigurationProperties.getWebEndpoint().getPrototypeFilter()
-      .getPath();
+    final String prototypeFilterPath =
+        spectatorStackdriverConfigurationProperties.getWebEndpoint().getPrototypeFilter().getPath();
 
     if (!prototypeFilterPath.isEmpty()) {
       log.error("Ignoring prototypeFilterPath because it is not yet supported.");
       measurementFilter = null;
       log.info("Configuring stackdriver filter from {}", prototypeFilterPath);
-      measurementFilter = PrototypeMeasurementFilter.loadFromPath(prototypeFilterPath).and(filterNotSpring);
+      measurementFilter =
+          PrototypeMeasurementFilter.loadFromPath(prototypeFilterPath).and(filterNotSpring);
     } else {
       measurementFilter = filterNotSpring;
     }
@@ -191,21 +194,33 @@ public class StackdriverConfig {
     String host = hostaddr.getCanonicalHostName();
     String hostPort = host + ":" + serverProperties.getPort();
 
-    ConfigParams params = new ConfigParams.Builder().setCounterStartTime(new Date().getTime()).setCustomTypeNamespace(
-      "spinnaker"
-    ).setProjectName(spectatorStackdriverConfigurationProperties.getStackdriver().getProjectName()).setApplicationName(
-      spectatorStackdriverConfigurationProperties.getApplicationName(environment.getProperty("spring.application.name"))
-    ).setCredentialsPath(spectatorStackdriverConfigurationProperties.getStackdriver().getCredentialsPath())
-      .setMeasurementFilter(measurementFilter).setInstanceId(hostPort).build();
+    ConfigParams params =
+        new ConfigParams.Builder()
+            .setCounterStartTime(new Date().getTime())
+            .setCustomTypeNamespace("spinnaker")
+            .setProjectName(
+                spectatorStackdriverConfigurationProperties.getStackdriver().getProjectName())
+            .setApplicationName(
+                spectatorStackdriverConfigurationProperties.getApplicationName(
+                    environment.getProperty("spring.application.name")))
+            .setCredentialsPath(
+                spectatorStackdriverConfigurationProperties.getStackdriver().getCredentialsPath())
+            .setMeasurementFilter(measurementFilter)
+            .setInstanceId(hostPort)
+            .build();
 
     stackdriver = new StackdriverWriter(params);
 
     Scheduler scheduler = Schedulers.from(Executors.newFixedThreadPool(1));
 
-    Observable.timer(spectatorStackdriverConfigurationProperties.getStackdriver().getPeriod(), TimeUnit.SECONDS)
-      .repeat().subscribe(interval -> {
-        stackdriver.writeRegistry(registry);
-      });
+    Observable.timer(
+            spectatorStackdriverConfigurationProperties.getStackdriver().getPeriod(),
+            TimeUnit.SECONDS)
+        .repeat()
+        .subscribe(
+            interval -> {
+              stackdriver.writeRegistry(registry);
+            });
 
     return stackdriver;
   }
