@@ -16,12 +16,14 @@
 
 package com.netflix.spinnaker.kork.telemetry.caffeine;
 
+import com.github.benmanes.caffeine.cache.RemovalCause;
 import com.github.benmanes.caffeine.cache.stats.CacheStats;
 import com.github.benmanes.caffeine.cache.stats.StatsCounter;
 import com.netflix.spectator.api.Counter;
 import com.netflix.spectator.api.Registry;
 import com.netflix.spectator.api.Timer;
 import java.util.concurrent.TimeUnit;
+import org.checkerframework.checker.index.qual.NonNegative;
 
 public class CaffeineStatsCounter implements StatsCounter {
   private final Counter hitCount;
@@ -71,12 +73,15 @@ public class CaffeineStatsCounter implements StatsCounter {
   }
 
   @Override
+  public void recordEviction(@NonNegative int weight, RemovalCause cause) {}
+
+  // @Override
   @SuppressWarnings("deprecation")
   public void recordEviction() {
     recordEviction(1);
   }
 
-  @Override
+  // @Override
   public void recordEviction(int weight) {
     evictionCount.increment();
     evictionWeight.increment(weight);
@@ -84,7 +89,7 @@ public class CaffeineStatsCounter implements StatsCounter {
 
   @Override
   public CacheStats snapshot() {
-    return new CacheStats(
+    return CacheStats.of(
         hitCount.count(),
         missCount.count(),
         loadSuccessCount.count(),
