@@ -12,23 +12,25 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT, classes = {
-  WebEnvironmentTest.TestControllerConfiguration.class})
+@SpringBootTest(
+    webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT,
+    classes = {WebEnvironmentTest.TestControllerConfiguration.class})
 // It would be lovely to have a per-method TestPropertySource annotation, or
 // some other simple way to parametrize tests to verify what happens when
 // default.rejectIllegalHeader isn't set at all, and set to true, in addition to
 // setting it to false. At the moment this doesn't seem worth the code
 // duplication / complexity. See
 // https://github.com/spring-projects/spring-framework/issues/18951.
-@TestPropertySource(properties = {"logging.level.org.apache.coyote.http11.Http11InputBuffer = DEBUG",
-  "default.rejectIllegalHeader = false"})
+@TestPropertySource(
+    properties = {
+      "logging.level.org.apache.coyote.http11.Http11InputBuffer = DEBUG",
+      "default.rejectIllegalHeader = false"
+    })
 class WebEnvironmentTest {
 
-  @LocalServerPort
-  int port;
+  @LocalServerPort int port;
 
-  @Autowired
-  TestRestTemplate restTemplate;
+  @Autowired TestRestTemplate restTemplate;
 
   /*
    * @Test void testTomcatWithIllegalHttpHeaders() throws Exception { HttpHeaders headers = new
